% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_check.R
\name{arg_check_mle}
\alias{arg_check_mle}
\title{Check the MLE function args}
\usage{
arg_check_mle(data, dmetric, optimization)
}
\arguments{
\item{data}{A list of x vector (x-dim), y vector (y-dim), and z observation vector}

\item{dmetric}{A string -  distance metric - "euclidean" or "great_circle"}

\item{optimization}{A list of opt lb values (clb), opt ub values (cub), tol, max_iters}
}
\value{
dmetric as integer
}
\description{
Check the MLE function args
}
