% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_check.R
\name{arg_check_predict}
\alias{arg_check_predict}
\title{Check the prediction function args}
\usage{
arg_check_predict(data_train, data_test, theta, dmetric)
}
\arguments{
\item{data_train}{A list of x vector (x-dim), y vector (y-dim), and z observation vector}

\item{data_test}{A list of x vector (x-dim) and y vector (y-dim)}

\item{dmetric}{A string -  distance metric - "euclidean" or "great_circle"}

\item{theta:}{list of n parameters}
}
\value{
dmetric as integer
}
\description{
Check the prediction function args
}
