% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm.R
\name{bm}
\alias{bm}
\title{Benchmark function to run a given FUN on a set of synthetic datasets generated by ExaGeoStatR}
\usage{
bm(
  FUN,
  dmetric = "euclidean",
  n = 400,
  min_seed = 0,
  max_seed = 9,
  ts = 320,
  lts = 0,
  ncores = 4,
  ngpus = 0,
  pgrid = 1,
  qgrid = 1
)
}
\arguments{
\item{FUN:}{A predefined function to perform both modeling and prediction operations}

\item{dmetric:}{A string -  distance metric - "euclidean" or "great_circle"}

\item{n:}{An integer -  numer of spatial locations}

\item{min_seed:}{An integer -  initial seed to generate the synthetic datasets}

\item{max_seed:}{An integer -  last seed to generate the synthetic datasets}

\item{ncores:}{An integer -  numer of CPU cores to use}

\item{ngpus:}{An integer -  numer of GPUs to use}
}
\value{
a list of prediction errors (MSE_vec, RMSE_vec, MAE_vec, MSLE_vec)
}
\description{
Benchmark function to run a given FUN on a set of synthetic datasets generated by ExaGeoStatR
}
