% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_comp.R
\name{bm_comp}
\alias{bm_comp}
\title{Running KAUST 2021 competition using the Benchmark function 
Benchmark function to run a given FUN on KAUST 2021 competition datasets generated by ExaGeoStatR}
\usage{
bm_comp(FUN, Data_train_list, Data_predict_list)
}
\arguments{
\item{FUN:}{A predefined function to perform both modeling and prediction operations}

\item{Data_train_list:}{A list of trainig data}

\item{Data_predict_list:}{A list of predict data}
}
\value{
a list of prediction errors (MS, RMSE, MAE, MSLE) and the estimate of the parameters (sigma, beta, nu, nuggets)
}
\description{
Running KAUST 2021 competition using the Benchmark function 
Benchmark function to run a given FUN on KAUST 2021 competition datasets generated by ExaGeoStatR
}
