% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_check.R
\name{check_dmetric}
\alias{check_dmetric}
\title{Check the distance metric input to be "euclidean" or "great_circle"}
\usage{
check_dmetric(dmetric)
}
\arguments{
\item{dmetric:}{string}
}
\value{
dmetric as integer
}
\description{
Check the distance metric input to be "euclidean" or "great_circle"
}
