% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_check.R
\name{check_kernel}
\alias{check_kernel}
\title{Check the statistical kernel to be in ("ugsm-s", "ugsmn-s",  "bgsfm-s", "bgspm-s", "tgspm-s", "ugsm-st", "bgsm-st")}
\usage{
check_kernel(kernel)
}
\arguments{
\item{kernel:}{string}
}
\value{
kernel as integer
}
\description{
Check the statistical kernel to be in ("ugsm-s", "ugsmn-s",  "bgsfm-s", "bgspm-s", "tgspm-s", "ugsm-st", "bgsm-st")
}
