% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_check.R
\name{check_theta}
\alias{check_theta}
\title{Check the statistical parameter vector (theta)}
\usage{
check_theta(theta)
}
\arguments{
\item{theta:}{list of n parameters}
}
\value{
N/A
}
\description{
Check the statistical parameter vector (theta)
}
