% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exageostat_test_wrapper.R
\name{dst_mle}
\alias{dst_mle}
\title{Maximum Likelihood Evaluation (MLE) using Diagonal Super-tile (DST) method}
\usage{
dst_mle(data = list(x, y, z), dst_thick, dmetric = 0,
  optimization = list(clb = c(0.001, 0.001, 0.001), cub = c(5, 5, 5), tol
  = 1e-04, max_iters = 100))
}
\arguments{
\item{data}{A list of x vector (x-dim), y vector (y-dim), and z observation vector}

\item{dst_thick}{A number - Diagonal Super-Tile (DST) diagonal thick}

\item{dmetric}{A string -  distance metric - "dm" or "gcd"}

\item{optimization}{A list of opt lb (clb), opt ub (cub), tol, max_iters}
}
\value{
vector of three values (theta1, theta2, theta3)
}
\description{
Maximum Likelihood Evaluation (MLE) using Diagonal Super-tile (DST) method
}
