% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exageostat_test_wrapper.R
\name{exact_mle}
\alias{exact_mle}
\title{Maximum Likelihood Evaluation  using exact method}
\usage{
exact_mle(data = list(x, y, z), dmetric = 0, optimization = list(clb
  = c(0.001, 0.001, 0.001), cub = c(5, 5, 5), tol = 1e-04, max_iters =
  100))
}
\arguments{
\item{data}{A list of x vector (x-dim), y vector (y-dim), and z observation vector}

\item{dmetric}{A string -  distance metric - "dm" or "gcd"}

\item{optimization}{A list of opt lb values (clb), opt ub values (cub), tol, max_iters}
}
\value{
vector of three values (theta1, theta2, theta3)
}
\description{
Maximum Likelihood Evaluation  using exact method
}
