% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_mloe_mmom.R
\name{exact_mloe_mmom}
\alias{exact_mloe_mmom}
\title{Mean Misspecification of the Mean Square Error (MMOM) and Mean Loss of Efficiency (MLOE) using exact method}
\usage{
exact_mloe_mmom(
  data = list(x_train, y_train, z_train, x_test, y_test),
  kernel = c("ugsm-s", "ugsmn-s", "ugnsm-s", "bgsfm-s", "bgsbm-s", "ugsm-s", "ugsm-st"),
  dmetric = c("euclidean", "great_circle"),
  est_theta,
  true_theta,
  computation = 0
)
}
\arguments{
\item{dmetric}{string -  distance metric - "euclidean" or "great_circle"}

\item{data:}{list of training and testing vectors}

\item{kernel:}{string - kernel  ("ugsm-s", "ugsmn-s",  "bgsfm-s", "bgspm-s", "tgspm-s", "ugsm-st", "bgsm-st")}

\item{est_theta:}{list of n parameters (estimated theta)}

\item{true_theta:}{list of n parameters (true theta)}

\item{computation:}{integer -  should be always dense}
}
\value{
vector of MLOE/MMOM values
}
\description{
Mean Misspecification of the Mean Square Error (MMOM) and Mean Loss of Efficiency (MLOE) using exact method
}
