% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_predict.R
\name{exact_predict}
\alias{exact_predict}
\title{Perform prediction on testing data using training data and pre-estimated theta vector}
\usage{
exact_predict(
  data_train = list(x, y, z),
  data_test = list(x, y),
  kernel = c("ugsm-s", "ugsmn-s", "ugnsm-s", "bgsfm-s", "bgsbm-s", "ugsm-s", "ugsm-st"),
  dmetric = c("euclidean", "great_circle"),
  theta,
  computation = 0
)
}
\arguments{
\item{dmetric}{string -  distance metric - "euclidean" or "great_circle"}

\item{data_train:}{list of training data}

\item{data_test:}{list of testing data}

\item{kernel:}{string - kernel  ("ugsm-s", "ugsmn-s",  "bgsfm-s", "bgspm-s", "tgspm-s", "ugsm-st", "bgsm-st")}

\item{theta:}{list of n parameters (estimated theta)}

\item{computation:}{integer -  computation method}
}
\value{
list of predicted values
}
\description{
Perform prediction on testing data using training data and pre-estimated theta vector
}
