\name{exageostat_finalizeR}
\alias{exageostat_finalizeR}
\title{ Sets up the call to the rexageostat_finalize() C language function accessed via the  R '.C' external function call API }
\description{
  \code{exageostat_finalize} demonstrates the R / C language interface
}

\usage{
exageostat_finalizeR()
}

\arguments{
}

\value{
   Returned value: None

}
\author{ Sameh Abdulah, Postdoctoral Fellow, Extreme Computing and Research Center, KAUST \email{sameh.abdulah@kaust.edu.sa}}

\examples{
#
# inline call at R command line or in another R function:
# Provides function arguments for and calls the R exageostat__gen_zR() function.
#
#exageostat_finalize()
}


\keyword{programming}
