% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_finalize.R
\name{exageostat_init}
\alias{exageostat_init}
\title{Initial an instance of ExaGeoStatR}
\usage{
exageostat_init(
  hardware = list(ncores = 2, ngpus = 0, ts = 320, lts = 0, pgrid = 1, qgrid = 1)
)
}
\arguments{
\item{hardware}{A list of ncores, ngpus, tile size, pgrid, and qgrid}
}
\value{
N/A
}
\description{
Initial an instance of ExaGeoStatR
}
\examples{
exageostat_init(hardware = list(ncores = 2, ngpus = 0, ts = 320, lts = 0, pgrid = 1, qgrid = 1))
exageostat_init(hardware = list(ncores = 1, ngpus = 2, ts = 320, lts = 0, pgrid = 1, qgrid = 1))
exageostat_init(hardware = list(ncores = 26, ngpus = 0, ts = 320, lts = 0, pgrid = 3, qgrid = 4))
}
