% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_info.R
\name{fisher_general}
\alias{fisher_general}
\title{Compute the Fisher information matrix for a given data and theta vector}
\usage{
fisher_general(data = list(x, y), theta, dmetric)
}
\arguments{
\item{dmetric}{string -  distance metric - "euclidean" or "great_circle"}

\item{data:}{list of data vectors, x, and y}

\item{theta:}{list of n parameters (estimated theta)}
}
\value{
list of fisher matrix elements
}
\description{
Compute the Fisher information matrix for a given data and theta vector
}
