% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_absolute_error.R
\name{mean_absolute_error}
\alias{mean_absolute_error}
\title{Mean Absolute Error used as an assessment tool}
\usage{
mean_absolute_error(y, ypre)
}
\arguments{
\item{y<-}{c(5:20) - vector representing  number of true values}

\item{ypre<-}{predict(lm(y ~ x)) - vector denoting values of number of y predicted values.}
}
\value{
MAE is the average of the absolute error
}
\description{
Mean Absolute Error used as an assessment tool
}
