% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_square_error.R
\name{mean_square_error}
\alias{mean_square_error}
\title{Mean Square Error used as an assessment tool}
\usage{
mean_square_error(y, ypre)
}
\arguments{
\item{y<-}{c(5:20) - vector representing  number of true values}

\item{ypre<-}{predict(lm(y ~ x)) - vector denoting values of number of y predicted values.}
}
\value{
MSE is the is the average squares of the “errors”
}
\description{
Mean Square Error used as an assessment tool
}
