% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_squared_logarithmic_error.R
\name{mean_squared_logarithmic_error}
\alias{mean_squared_logarithmic_error}
\title{Mean Absolute Error used as an assessment tool}
\usage{
mean_squared_logarithmic_error(y, ypre)
}
\arguments{
\item{y<-}{c(5:20) - vector representing  number of true values}

\item{ypre<-}{predict(lm(y ~ x)) - vector denoting values of number of y predicted values}
}
\value{
MSLE is measure of the ratio between the true and predicted values
}
\description{
Mean Absolute Error used as an assessment tool
}
