% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.kriging.R
\name{plot.Krig}
\alias{plot.Krig}
\title{This function plots the the diagnostics and summaries of kriging
one thing to note here is that x is Krig or sreg object. need to figure out x."plot.Krig" <- function(x, digits = 4, which = 1:4,}
\usage{
\method{plot}{Krig}(x, digits = 4, which = 1:4, ...)
}
\description{
This function plots the the diagnostics and summaries of kriging
one thing to note here is that x is Krig or sreg object. need to figure out x."plot.Krig" <- function(x, digits = 4, which = 1:4,
}
