% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_mean_squared_error.R
\name{root_mean_squared_error}
\alias{root_mean_squared_error}
\title{Root Mean Squared Error used as an assessment tool}
\usage{
root_mean_squared_error(y, ypre)
}
\arguments{
\item{y<-}{c(5:20) - vector representing  number of true values}

\item{ypre<-}{predict(lm(y ~ x)) - vector denoting values of number of y predicted values.}
}
\value{
RMSE is the square root of the mean of the square of all of the error
}
\description{
Root Mean Squared Error used as an assessment tool
}
