% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exageostat_test_wrapper.R
\name{simulate_data_exact}
\alias{simulate_data_exact}
\title{Simulate Geospatial data (x, y, z)}
\usage{
simulate_data_exact(sigma_sq, beta, nu, dmetric = 0, n, seed = 0)
}
\arguments{
\item{sigma_sq}{A number - variance parameter}

\item{beta}{A number - smoothness parameter)}

\item{nu}{A number  - range parameter}

\item{dmetric}{A string -  distance metric - "dm" or "gcd"}

\item{n}{A number -  data size}

\item{seed}{A number -  seed of random generation}
}
\value{
a list of of three vectors (x, y, z)
}
\description{
Simulate Geospatial data (x, y, z)
}
\examples{
data = simulate_data_exact( sigma_sq, beta, nu, dmetric, seed)
}
