% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exageostat_test_wrapper.R
\name{simulate_obs_exact}
\alias{simulate_obs_exact}
\title{Simulate Geospatial data given (x, y) locations}
\usage{
simulate_obs_exact(x, y, sigma_sq, beta, nu, dmetric = 0)
}
\arguments{
\item{x}{A vector  (x-dim)}

\item{y}{A vector (y-dim)}

\item{sigma_sq}{A number - variance parameter}

\item{beta}{A number - smoothness parameter)}

\item{nu}{A number  - range parameter}

\item{dmetric}{A string -  distance metric - "dm" or "gcd"}
}
\value{
a list of of three vectors (x, y, z)
}
\description{
Simulate Geospatial data given (x, y) locations
}
\examples{
data = simulate_obs_exact(x, y, sigma_sq, beta, nu, dmetric)
}
