% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_data.R
\name{splitting_data}
\alias{splitting_data}
\title{Spliting data into training and testing datasets}
\usage{
splitting_data(Datatray, k = 10, n = 400)
}
\arguments{
\item{Datatray:}{the full dataset}

\item{k:}{testing dataset portion k\%}

\item{n:}{number of spatial locations}
}
\value{
a list of training and testing datasets
}
\description{
Spliting data into training and testing datasets
}
