% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exageostat_test_wrapper.R
\name{Test1}
\alias{Test1}
\title{Test Generating Z vector using random (x, y) locations with exact MLE computation.}
\usage{
Test1()
}
\description{
Test Generating Z vector using random (x, y) locations with exact MLE computation.
}
\examples{
Test1()
}
