% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exageostat_test_wrapper.R
\name{Test2}
\alias{Test2}
\title{Test Generating Z vector using random (x, y) locations with TLR MLE computation.}
\usage{
Test2()
}
\description{
Test Generating Z vector using random (x, y) locations with TLR MLE computation.
}
\examples{
Test2()
}
