% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exageostat_test_wrapper.R
\name{Test3}
\alias{Test3}
\title{Test Generating Z vector using random (x, y) locations with DST MLE computation.}
\usage{
Test3()
}
\description{
Test Generating Z vector using random (x, y) locations with DST MLE computation.
}
\examples{
Test3()
}
