% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exageostat_test_wrapper.R
\name{Test4}
\alias{Test4}
\title{Test Generating Z vector using given (x, y) locations with exact MLE computation.}
\usage{
Test4()
}
\description{
Test Generating Z vector using given (x, y) locations with exact MLE computation.
}
\examples{
Test4()
}
