\name{TestWrapper}
\alias{TestWrapper}
\title{ Test Wrapper Function for TrialSwapR interface to the CalcMatAvgR() function  }
\description{
  \code{TestWrapper} demonstrates the R / C language interface
}

\usage{
TestWrapper()
}

\arguments{ None. }

\value{ Returns nothing. }


\author{ Rick Reeves,Scientific Programmer, National Center for Ecological Analysis and Synthesis, UCSB \email{reeves@nceas.ucsb.edu}}

\examples{
#
# inline call at R command line or in another R function:
# Provides function arguments for and calls the R CalcMatAvgR() function.
#
#   TestWrapper()
}
\keyword{programming}

