% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exageostat_test_wrapper.R
\name{exageostat_finalize}
\alias{exageostat_finalize}
\title{Finalize the current instance of ExaGeoStatR}
\usage{
exageostat_finalize()
}
\value{
N/A
}
\description{
Finalize the current instance of ExaGeoStatR
}
\examples{
exageostat_finalize()
}
