% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data_exact}
\alias{simulate_data_exact}
\title{Simulate Geospatial data (x, y, z)}
\usage{
simulate_data_exact(
  sigma_sq,
  beta,
  nu,
  dmetric = c("euclidean", "great_circle"),
  n,
  seed = 0
)
}
\arguments{
\item{sigma_sq}{A number - variance parameter}

\item{beta}{A number - smoothness parameter)}

\item{nu}{A number  - range parameter}

\item{dmetric}{A string -  distance metric - "euclidean" or "great_circle"}

\item{n}{A number -  data size}

\item{seed}{A number -  seed of random generation}
}
\value{
a list of of three vectors (x, y, z)
}
\description{
Simulate Geospatial data (x, y, z)
}
\examples{
seed = 0  ##Initial seed to generate XY locs.
sigma_sq = 1 ##Initial variance.
beta = 0.1 ##Initial smoothness.
nu = 0.5 ##Initial range.
dmetric = "euclidean" ##"euclidean" or "great_circle" distance.
n = 1600 ## The number of locations (n must be a square number, n=m^2).
exageostat_init(hardware = list (ncores = 2, ngpus = 0,
                ts = 320, pgrid  = 1, qgrid  = 1)) ##Initiate exageostat instance
data = simulate_data_exact(sigma_sq, beta, nu, dmetric,
                           n, seed) ##Generate Z observation vector
data
exageostat_finalize() ##Finalize exageostat instance
}
