% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_xy.R
\name{simulate_obs_exact}
\alias{simulate_obs_exact}
\title{Simulate Geospatial data given (x, y) locations}
\usage{
simulate_obs_exact(
  x,
  y,
  sigma_sq,
  beta,
  nu,
  dmetric = c("euclidean", "great_circle")
)
}
\arguments{
\item{x}{A vector  (x-dim)}

\item{y}{A vector (y-dim)}

\item{sigma_sq}{A number - variance parameter}

\item{beta}{A number - smoothness parameter)}

\item{nu}{A number  - range parameter}

\item{dmetric}{A string -  distance metric - "euclidean" or "great_circle"}
}
\value{
a list of of three vectors (x, y, z)
}
\description{
Simulate Geospatial data given (x, y) locations
}
\examples{
sigma_sq = 1 ##Initial variance.
beta = 0.1 ##Initial smoothness.
nu = 0.5 ##Initial range.
dmetric = "euclidean" ##"euclidean" or "great_circle" distance.
n = 1600 ## The number of locations (n must be a square number, n=m^2)
x = rnorm(n, 0, 1)     #x measurements of n locations.
y = rnorm(n, 0, 1)    #y measurements of n locations.
exageostat_init(hardware = list (ncores = 2, ngpus = 0, ts = 320,
                pgrid  = 1, qgrid  = 1)) ##Initiate exageostat instance
data = simulate_obs_exact(x, y, sigma_sq, beta, nu,
                 dmetric) ##Generate Z observation vector based on given locations
exageostat_finalize() ##Finalize exageostat instance
}
